==> 02_group_measurements.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 16, 8

tr = pd.read_csv('input/train.csv')

te = pd.read_csv('input/test.csv')
te['Expected'] = -1
te['Id'] += 10000000

d = pd.concat([tr, te])
d.index = pd.Index(range(len(d)))

tr = None
te = None

grouped = d.groupby('Id')

d = None

def aggregate_groups(group):
    null_ratio = (1.0 * group.isnull().sum().sum()) / (len(group) * (len(group.columns) - 2))
    vals = [
        group['Id'].min(),
        group['Expected'].mean(),
        len(group),
        null_ratio
    ]
    cols = [
        'Id',
        'Expected',
        'cnt',
        'null_ratio'
    ]
    for c in group.columns:
        if c not in ['Expected', 'Id']:
            vals.append(group[c].mean())
            cols.append(c + '_mean')
            vals.append(group[c].min())
            cols.append(c + '_min')
            vals.append(group[c].max())
            cols.append(c + '_max')
            vals.append(group[c].std())
            cols.append(c + '_std')
    id = group['Id'].min()
    if id % 10000 == 0:
        print id
    return Series(vals, index=cols)

g = grouped.apply(aggregate_groups)

grouped = None

g.to_csv('tmp/grouped.csv', index=False)
