==> 04_feature_engineering.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 12, 6
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.ensemble import GradientBoostingRegressor

s = pd.read_csv('tmp/grouped.csv').sample(100000)

tr = s[s.Expected != -1]
te = s[s.Expected == -1]

tr.info()

base_metrics = [
    'Ref',
    'RefComposite',
    'RhoHV',
    'Zdr',
    'Kdp'
]
metric_modifiers = [
    '',
    '_5x5_10th',
    '_5x5_50th',
    '_5x5_90th'
]
metric_stats = [
    '_mean',
    '_min',
    '_max'
]
for bm1 in base_metrics:
    for mm1 in metric_modifiers:
        for ms1 in metric_stats:
            for bm2 in base_metrics:
                for mm2 in metric_modifiers:
                    for ms2 in metric_stats:
                        m1 = bm1 + mm1 + ms1
                        m2 = bm2 + mm2 + ms2
                        tr[m1 + '_per_' + m2] = tr[m1] / tr[m2]

tr.info()

trl = tr.Expected
del tr['Expected']

tr = tr.replace([np.inf, -np.inf], 0)
tr = tr.fillna(0)

regr = GradientBoostingRegressor(n_estimators=300, max_depth=5, )

regr.fit(tr, trl)

importances = regr.feature_importances_
indices = np.argsort(importances)[::-1]
for f in range(100):
    print("%d. feature %d (%f) %s" % (f + 1, indices[f], importances[indices[f]], tr.columns[indices[f]]))
