==> 05_feature_engineering.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt

d = pd.read_csv('tmp/grouped.csv')

important_features = []
with open('important_features.txt') as f:
    important_features = f.readlines()
important_features = [f.strip() for f in important_features][:55]
print important_features

base_metrics = [
    'Ref',
    'RefComposite',
    'RhoHV',
    'Zdr',
    'Kdp'
]
metric_modifiers = [
    '',
    '_5x5_10th',
    '_5x5_50th',
    '_5x5_90th'
]
metric_stats = [
    '_mean',
    '_min',
    '_max'
]
for bm1 in base_metrics:
    for mm1 in metric_modifiers:
        for ms1 in metric_stats:
            for bm2 in base_metrics:
                for mm2 in metric_modifiers:
                    for ms2 in metric_stats:
                        m1 = bm1 + mm1 + ms1
                        m2 = bm2 + mm2 + ms2
                        new_metric = m1 + '_per_' + m2
                        if new_metric in important_features:
                            d[new_metric] = d[m1] / d[m2]

print d.columns.tolist()

for c in d.columns[:92]:
    if c.endswith('_min') or c.endswith('_max'):
        del d[c]

print len(d.columns.tolist())

d = d.replace([np.inf, -np.inf], 0)
d = d.fillna(0)

d.to_csv('tmp/engineered_features_01.csv', index=False)
