==> 07_predict.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 12, 6
import xgboost as xgb
from sklearn.metrics import mean_absolute_error

d = pd.read_csv('tmp/engineered_features_01.csv')

d.Id = d.Id.astype(int)

tr = d[d.Expected > -1]
te = d[d.Expected == -1]

del d
trl = tr.Expected
del tr['Expected']
tel = te.Expected
del te['Expected']

m = xgb.XGBRegressor(n_estimators=500, nthread=6, max_depth=15, objective="reg:linear")
m.fit(tr, trl)

preds = m.predict(te)

#te.Id = te.Id - 10000000

sub = DataFrame({'Id': te.Id})
sub['Expected'] = preds
sub.to_csv('output/s_01.csv', index=False)
