==> 08_group_measurements_mp.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 16, 8

tr = pd.read_csv('input/train.csv',
                 usecols=['Id',
                          'minutes_past',
                          'radardist_km',
                          'Ref',
                          'Ref_5x5_50th',
                          'RefComposite',
                          'RefComposite_5x5_50th',
                          'Expected'])

te = pd.read_csv('input/test.csv',
                 usecols=['Id',
                          'minutes_past',
                          'radardist_km',
                          'Ref',
                          'Ref_5x5_50th',
                          'RefComposite',
                          'RefComposite_5x5_50th'])
te['Expected'] = -1
te['Id'] += 10000000

d = pd.concat([tr, te])
d.index = pd.Index(range(len(d)))
d.info()

del tr
del te

grouped = d.groupby('Id')

del d

def compute(g, c):
    valid_time = np.zeros_like(g.minutes_past)
    valid_time[0] = g.minutes_past.iloc[0]
    for n in xrange(1, len(g.minutes_past)):
        valid_time[n] = g.minutes_past.iloc[n] - g.minutes_past.iloc[n-1]
    valid_time[-1] = valid_time[-1] + 60 - np.sum(valid_time)
    valid_time = valid_time / 60.0
    
    sum_mp = 0
    sum_ecs = 0
    sum_wcs = 0
    sum_w8c = 0
    sum_rt = 0
    for dbz, hours in zip(g[c], valid_time):
        if np.isfinite(dbz):
            mp_perhr = pow(pow(10, dbz/10)/200, 0.625)
            sum_mp = sum_mp + mp_perhr * hours
            ecs_perhr = pow(pow(10, dbz/10)/130, 0.5)
            sum_ecs = sum_ecs + ecs_perhr * hours
            wcs_perhr = pow(pow(10, dbz/10)/75, 0.5)
            sum_wcs = sum_wcs + wcs_perhr * hours
            w8c_perhr = pow(pow(10, dbz/10)/300, 0.714)
            sum_w8c = sum_w8c + w8c_perhr * hours
            rt_perhr = pow(pow(10, dbz/10)/250, 0.833)
            sum_rt = sum_rt + rt_perhr * hours
    return (sum_mp, sum_ecs, sum_wcs, sum_w8c, sum_rt)

def aggregate_groups(group):
    group = group.sort('minutes_past', ascending=True)
    null_ratio = (1.0 * group.isnull().sum().sum()) / (len(group) * (len(group.columns) - 2))
    vals = [
        group.iloc[0]['Id'],
        group['Expected'].mean(),
        len(group),
        null_ratio,
        group['radardist_km'].mean(),
    ]
    cols = [
        'Id',
        'Expected',
        'cnt',
        'null_ratio',
        'radardist_km'
    ]
    radar_cols = ['Ref', 'Ref_5x5_50th', 'RefComposite', 'RefComposite_5x5_50th']
    for c in radar_cols:
        mp, ecs, wcs, w8c, rt = compute(group, c)
        vals.append(mp)
        cols.append(c + '_mp')
        vals.append(ecs)
        cols.append(c + '_ecs')
        vals.append(wcs)
        cols.append(c + '_wcs')
        vals.append(w8c)
        cols.append(c + '_w8c')
        vals.append(rt)
        cols.append(c + '_rt')
    if group['Id'].mean() % 100000 == 0:
        print group['Id'].mean()
    return Series(vals, index=cols)

g = grouped.apply(aggregate_groups)

del grouped

g.Id = g.Id.astype(int)
g.cnt = g.cnt.astype(int)

g.info()

g[:30]
g.to_csv('tmp/grouped_mp_02.csv')
