==> 10_predict.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 12, 6
import xgboost as xgb
from sklearn.metrics import mean_absolute_error

d = pd.read_csv('tmp/grouped_mp_02.csv')

e = d[(d.cnt < 10) &
           (d.null_ratio < 0.02)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.06, min_child_weight=1, n_estimators=1, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p1 = te[['Id', 'Expected']]
p1.describe()

e = d[(10 <= d.cnt) & (d.cnt < 15) &
           (d.null_ratio < 0.02)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.03, min_child_weight=1, n_estimators=50, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p2 = te[['Id', 'Expected']]
p2.describe()

e = d[(15 <= d.cnt) &
           (d.null_ratio < 0.02)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.01, min_child_weight=1, n_estimators=10, max_depth=5)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p3 = te[['Id', 'Expected']]
p3.describe()

e = d[(d.cnt < 10) &
           ( 0.02 <= d.null_ratio) & (d.null_ratio < 0.65)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.06, min_child_weight=1, n_estimators=50, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p4 = te[['Id', 'Expected']]
p4.describe()

e = d[(10 <= d.cnt) & (d.cnt < 15) &
           ( 0.02 <= d.null_ratio) & (d.null_ratio < 0.65)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.06, min_child_weight=1, n_estimators=50, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p5 = te[['Id', 'Expected']]
p5.describe()

e = d[(15 <= d.cnt) &
           ( 0.02 <= d.null_ratio) & (d.null_ratio < 0.65)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.01, min_child_weight=1, n_estimators=50, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p6 = te[['Id', 'Expected']]
p6.describe()

e = d[(d.cnt < 10) &
           ( 0.65 <= d.null_ratio)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.01, min_child_weight=1, n_estimators=10, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p7 = te[['Id', 'Expected']]
p7.describe()

e = d[(10 <= d.cnt) & (d.cnt < 15) &
           ( 0.65 <= d.null_ratio)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.01, min_child_weight=1, n_estimators=1, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p8 = te[['Id', 'Expected']]
p8.describe()

e = d[(15 <= d.cnt) &
           ( 0.65 <= d.null_ratio)]
tr = e[e.Expected != -1]
trl = tr.Expected
del tr['Expected']
te = e[e.Expected == -1]
tel = te.Expected
del te['Expected']
m = xgb.XGBRegressor(learning_rate=0.01, min_child_weight=1, n_estimators=1, max_depth=20)
m.fit(tr, trl)
te['Expected'] = m.predict(te)
p9 = te[['Id', 'Expected']]
p9.describe()

p = pd.concat([p1, p2, p3, p4, p5, p6, p7, p8, p9])
p.info()

p.describe()

s = pd.read_csv('input/sample_solution.csv')
s.info()

print len(p[p.Expected > 500])
print len(p[p.Expected > 1000])
print len(p[p.Expected > 1500])
print len(p[p.Expected < 0])

p.loc[p.Expected < 0,'Expected'] = 0
print len(p[p.Expected < 0])
p.describe()

p = p.sort_index()

p.Id = p.Id - 10000000
p.info()

p.describe()

p[:20]

p.to_csv('output/s_07.csv', index=False)

p.loc[p.Expected > 100,'Expected'] = 100

p.to_csv('output/s_08.csv', index=False)
